[Refactor/#3] 컴포넌트 분리 수정 및 함수 파라미터 수정
